backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
gunicorn==23.0.0
mercadopago==2.2.3
python-decouple==3.8
celery==5.4.0
redis==5.2.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
psycopg2-binary==2.9.10